requirements.txt
requests>=2.25.1
pandas>=1.3.0
tabulate>=0.8.9
pyyaml>=6.0
numpy>=1.21.0
python-dotenv>=1.0.0 